Move Slack join event webhook to /api/slack-join
